refactor and cleanup organizational structure search
