--- next/backend/dtos/minio/upload-minio-file.dto.ts
@@ -1,6 +1,0 @@
-export interface UploadMinioFile {
-  file: File
-  originalName: string
-  errorMessage?: string
-  isUploading?: boolean
-}
