+++ next/backend/ms-graph/constants.ts
@@ -1,0 +1,4 @@
+/**
+ * 'Struktura pre web' group id
+ */
+export const MS_GRAPH_GROUP_ID = 'e2a318ff-fadb-4950-8ff0-a69660788e9d'
