+++ next/pages/api/ms-graph/structure.ts
@@ -1,12 +1,11 @@
-import { getGroupMembersRecursive, getToken } from 'backend/services/ms-graph'
+import { MS_GRAPH_GROUP_ID } from '@backend/ms-graph/constants'
+import { getMsalToken } from '@backend/ms-graph/getMsalToken'
+import { getGroupMembersRecursive } from '@backend/ms-graph/ms-graph'
 import type { NextApiRequest, NextApiResponse } from 'next'
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  const { accessToken } = await getToken()
-  // e2a318ff-fadb-4950-8ff0-a69660788e9d is id of 'Struktura pre web' group
-  return res.json(
-    await getGroupMembersRecursive(accessToken, 'e2a318ff-fadb-4950-8ff0-a69660788e9d', null),
-  )
+  const { accessToken } = (await getMsalToken()) ?? {}
+  return res.json(await getGroupMembersRecursive(accessToken ?? '', MS_GRAPH_GROUP_ID, null))
 }
 
 export default handler
