+++ next/utils/utils.ts
@@ -6,14 +6,6 @@ export const arrayify = (input: string | string[] | undefined | null) => {
   return input
 }
 
-// turn unknown values into (empty) strings, return string if it's a string
-export const forceString = (input: unknown) => {
-  if (typeof input === 'string') return input
-  if (typeof input === 'number') return input.toString()
-  if (Array.isArray(input)) return input.join(', ')
-  return ''
-}
-
 export const isPresent = <U>(a: U | null | undefined | void): a is U => {
   return a !== undefined && a !== null
 }
